Document `plugins` in the ApolloServer API reference.
This option was previously not explicitly documented except on the plugins
page. Add a basic description and mention that the built-in plugins enable
themselves in some contexts. (I didn't mention that the JSON tracing and
cache-control plugins work the same way; the fact that those features are
implemented as plugins is an internal implementation as those plugins are not
directly exposed to users.)

There are many other ApolloServer options not mentioned on this reference page!
In case you are curious, they are:

- cache
- cacheControl
- dataSources
- executor
- extensions
- fieldResolver
- gateway
- modules
- parseOptions
- tracing
- uploads